*** empty log message ***

+++ src/engine/sound.cpp
@@ -208,8 +208,8 @@ void updatechanvol(int chan, const vec *loc, int svol)
         if(vol<0) vol = 0;
         if(stereo && (v.x != 0 || v.y != 0))
         {
-            float yaw = -atan2(v.x, v.y) - camera1->yaw*RAD; // relative angle of sound along X-Y axis
-            pan = int(255.9f*(0.5f*sin(yaw)+0.5f)); // range is from 0 (left) to 255 (right)
+            float yaw = -atan2f(v.x, v.y) - camera1->yaw*RAD; // relative angle of sound along X-Y axis
+            pan = int(255.9f*(0.5f*sinf(yaw)+0.5f)); // range is from 0 (left) to 255 (right)
         };
     };
     vol = (vol*MAXVOL*svol)/255/255;
